add BTA_pd_matrices
