--- include/old-code/old_stacked_pd_linalg.h
@@ -1,52 +1,0 @@
-/*
- * Copyright 2026 Daniel Cederberg and William Zhang
- *
- * This file is part of the SparseDiffEngine project.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-#ifndef OLD_STACKED_PD_LINALG_H
-#define OLD_STACKED_PD_LINALG_H
-
-#include "utils/CSC_matrix.h"
-#include "utils/permuted_dense.h"
-#include "utils/stacked_pd.h"
-
-/* Legacy "B is a stacked_pd" BA kernels: C = B @ A where B is spd and A is
-   CSC / permuted_dense / stacked_pd. No production atom currently produces
-   a stacked_pd in the *left*-operand position of a B @ A composition
-   (stacked_pds appear only as derived Jacobians, never as constant
-   coefficients), so these kernels and their matrix_BTA dispatcher entries
-   were moved out of production paths. They are kept here as a reference
-   implementation and exercised by tests/old-code/test_old_stacked_pd.h. */
-
-/* Allocate C = B @ A where B is spd, A is CSC. */
-matrix *BA_spd_csc_alloc(const stacked_pd *B, const CSC_matrix *A);
-
-/* Fill values of C = B @ A where B is spd and A is CSC. */
-void BA_spd_csc_fill_values(const stacked_pd *B, const CSC_matrix *A, stacked_pd *C);
-
-/* Allocate sparsity for C = B @ A where B spd and A is a permuted_dense. */
-matrix *BA_spd_pd_alloc(const stacked_pd *B, const permuted_dense *A);
-
-/* Fill values of C = B @ A where B is pd and A is a permuted_dense. */
-void BA_spd_pd_fill_values(const stacked_pd *B, const permuted_dense *A,
-                           stacked_pd *C);
-
-/* Allocate sparsity for C = B @ A where both B and A are spds */
-matrix *BA_spd_spd_alloc(const stacked_pd *B, const stacked_pd *A);
-
-/* Fill values of C = B @ A where both B and A are spds */
-void BA_spd_spd_fill_values(const stacked_pd *B, const stacked_pd *A, stacked_pd *C);
-
-#endif /* OLD_STACKED_PD_LINALG_H */
